Fixed the type of frameWord

+++ dev/lib/rxio/MDPHeader.cpp
@@ -58,8 +58,8 @@ using gpstk::BinUtils::decodeVar;
 
 namespace gpstk
 {
-   const unsigned MDPHeader::myLength = 16;
-   const unsigned MDPHeader::frameWord = 0x9c9c;
+   const uint16_t MDPHeader::myLength = 16;
+   const uint16_t MDPHeader::frameWord = 0x9c9c;
 
    // Set to zero for no debugging output
    // set to 1 to output text messages about decode/format/range errors
